getAll

+++ src/main/java/pl/example/repository/VoivodeshipRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import pl.example.models.VoivodeshipEntity;
 
 @Repository
-public interface VoivodeshipRepository extends CrudRepository<VoivodeshipEntity, Long> {
+public interface VoivodeshipRepository extends CrudRepository<VoivodeshipEntity, Integer> {
 }
